+++ app/params/manage/organisation_params.rb
@@ -15,7 +15,8 @@ module Manage
          last_minute_warning upcoming_soon_window invoice_payment_deadline
          deposit_payment_deadline deadline_postponable_for payment_overdue_deadline
          occupied_occupancy_color tentative_occupancy_color closed_occupancy_color
-         begins_at_default_time ends_at_default_time default_calendar_view]
+         begins_at_default_time ends_at_default_time default_calendar_view] +
+        [{ occupied_occupancy_states: [] }]
     end
 
     def self.admin_permitted_keys