+++ spec/models/organisation_settings_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe OrganisationSettings, type: :model do
+  subject { described_class.new }
+end