+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Feature: Add soft delete for booking categories
 - Feature: Add custom booking questions
 - Fix: Implement QR bill standards more closely
+- Feature: Add definitive state options for organisation
 
 ## Version 23.6.2